+++ SW/visualisation/src/src/main/java/Interfaces/ILightable.java
@@ -1,4 +1,6 @@
-package Light;
+package Interfaces;
+
+import Light.TypeOfLight;
 
 /**
  * Created by Perec on 2016.03.06..
