+++ SW/visualisation/src/src/main/java/Interfaces/IMileage.java
@@ -1,4 +1,4 @@
-package HMI;
+package Interfaces;
 
 /**
  * Created by haxxi on 2016.03.04..
