+++ SW/visualisation/src/src/main/java/Interfaces/IWheelVisualization.java
@@ -1,0 +1,9 @@
+package Interfaces;
+
+/**
+ * Created by secured on 2016. 03. 13..
+ */
+public interface IWheelVisualization {
+    float ModifyVehicleSpeed(float speed);
+    float ModifyVehicleOrientaton(float degree);
+}
