+++ SW/visualisation/src/src/main/java/Listeners/OnVehicleListener.java
@@ -1,0 +1,10 @@
+package Listeners;
+
+import javax.swing.text.Position;
+
+/**
+ * Created by secured on 2016. 03. 13..
+ */
+public interface OnVehicleListener {
+    float PositionChanged(Position position);
+}
